requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-annoying==0.10.6
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.18
mysqlclient==1.4.6
oauthlib==3.1.0
protobuf==3.11.2
pycparser==2.19
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
